add optional priority shuffle to 3hop xn shuffle
xn_shuffle_init() takes a second arg that is the priority shuffle
callback function.  if it is NULL then the priority shuffle is
not enabled.  if it is not NULL, then a priority shuffle instance
is created (this shuffle does not batch RPCs, so they should be
sent as soon as they are posted).  data received on the priority
shuffle is delivered to the given priority shuffle callback (rather
than the main data callback).   to send data on the priority shuffle,
use the xn_shuffle_priority_send() API.   note that it is an error
to attempt to send priority shuffle data without enabling it (your
program will crash or assert in this case).

this code is a minorly revised version of the code from Ankush's
refactor-carp branch.