fix case where we directly do native write to ourselves
since we swapped the indexed prop for the filename and
put the filename in with the data the p.data_ptr should
be offset by the sizeof the indexed prop (currently hardwired
at sizeof(float)) plus the message type and the \0 it put
in the buffer.  the data buffer length is the fname_len+data_len.