realign code to match key/value store used with carp
At the application level, the app (e.g. vpic) uses the preloaded
stdio API to open a "filename" and write "filedata" to it (using
the file-per-particle model).   The preload library takes the filename
and filedata and transforms them into the preload "inkey" and "invalue"
(i.e. the key/value at the preload input).   The inkey and invalue
are then serialized for the shuffle (optionally adding "shuffle_extrabytes"
of zeros to allow you to test increasing the shuffle overhead without
increasing the storage costs).  So, the shuffle key/value will match
the preload inkey/invalue.

After the shuffle, the preload library transforms the preload
inkey and invalue into the preload "outkey" and "outvalue" (i.e.
the key/value at the preload output, before it goes to the deltafs
backend).

In the deltafs backend, the preload outkey and outvalue are transformed
to the key and value stored in the backend itself.

The code now uses these names for sanity and paranoid size checks.

Example values: the VPIC app has filename_size==19 and filedata_size==40
(both in bytes).

For the default mode: the preload inkey is the filename and the
preload invalue is the filedata (no transform required).   The
preload outkey/outvalue are the same as the inkey/invalue (no change
needed).   The backend key is a hash of the outkey (i.e. the filename)
and is typically 8 bytes.  The backend value is the the outvalue
(i.e. the filedata).

For the sideio mode (wisc): the filedata is written to a local log
at the log offset.   the preload inkey is the filename and the preload
invalue is the log offset (8 bytes).  the preload outkey is the same
as the inkey, but the preload outvalue is the source rank (4 bytes)
prepended to the 8 bytes offset in the invalue (so the outvalue is
12 bytes).  The backend key is a hash of the outkey (i.e. the filename)
and the backend value is the 12 byte outvalue (rank+offset).

For the sideft mode (bloomy): we write the key/value data locally
using deltafs first.   the preload inkey is the filename and
the preload invalue is empty.  so we are just shuffling the inkey
(i.e. the filename) without any data.  the preload outkey/outvalue
are the same as the inkey/invalue (no change needed).  At the backend
we write the outkey to a filter using deltafs_plfsdir_filter_put().

For carp range queries: the preload inkey is the attribute being
indexed (e.g. a 4 byte float that is extracted from the filedata).  the
preload invalue is the filename and filedata concatenated together
(19+40 bytes).  the preload outkey/outvalue are the same as the
inkey/invalue (no change needed).   The backend key/value is
the preload outkey/outvalue (no changes needed).

With this schemes there are 3 places where transforms can occur:
 1. filename/filedata => preload inkey/invalue  (in preload fclose())
 2. preload inkey/invalue => preload outkey/outvalue  (in preload_write())
 3. preload outkey/outvalue => deltafs key/value  (in deltafs lib code)

For this, we rename environment variables:
  PRELOAD_Particle_id_size => PRELOAD_Filename_size
  PRELOAD_Particle_size => PRELOAD_Filedata_size
  PRELOAD_Particle_extra_size => PRELOAD_Particle_extra_shuffle_bytes
(but we still honor the old names for backward compat)

We also change the shuffle serialize so that it does not add a
null between the shuffle key and value.  Modes that need this
null will include the trailing null in the key length instead.
(e.g. this is needed for cases where the shuffle key is a filename
string that is strlen()'d and hashed on the remote end).