remove redundant writes to Pivots pivots_ vector
We remove the redundant writes to pivots_ by aligning vector
management with what we do in OrderedBins.  At constructor
time we reserve() the pivots_ buffer without filling it
and we save the given pivot_count and later use it to sanity
check loading the pivots with data (e.g. the size you load
should match the size you reserved with the constructor).

no longer need is_set_ since we can tell if the Pivots have
been set by checking the size of pivots_.  If it is 0 then
it is !is_set, otherwise it must be set (to length pivot_count_).
we also no longer use Resize(), so we remove it.

Pivots::Size() now returns the desired pivot_count_ (even if
we have not loaded the pivots yet).

Fix err in Pivots::Calculate().  It should FillZeros() if
the combined consumer has zero total weight (was checking
if the Pivots::Size()==0 before, but that could not ever be
true since we just called FillZeros() before the check).

Fix issue this unconvered in pivot_bench where the pivot for
aggregation was setup with the wrong initial size (but got
adjusted to the correct value later).  Also removed redundant
calls to Pivots::FillZero() -- you should not have to fill
with zeros prior to calling load (since that will overwrite
the zeros you just wrote).