handle zero-weight pivots
Pivot::Calculate() will fill the pivots with zeros and set the weight
to 0 if there is no data.  This zero-filled pivot gets passed up the
RTP tree.  If one of the nodes up the tree receives zero-filled pivots
from all its children, then it will attempt to ComputeAggregatePivots()
on a set of zero-filled pivots.  Update ComputeAggregatePivots() to
catch this case early and just return a zero-filled pivot rather than
calling pivot_union() on it (pivot_union() will reject this case because
the list of bounds generated by LoadBounds() is empty because everything
is zero -- i.e. the entire range is zero-width).

Print a warning if the top-level proc decides to broadcast a zero-weight
set of pivots at the end of an RTP (there should be some weight, or else
why run RTP?).

Also, no need to copy the bounds list in the call to pivot_union()
so add a "&" ...