track in-bounds just using bins, not bins and OOB buffer
In-bounds/Out-of-bounds was being tracked in two places: first
by the boundaries of the OrderedBins and second by the ibrange_
in OobBuffer.  This is redundant and required apis to keep the
bounds defined by bins and OobBuffer in sync.   To resolve,
we remove ibrange_ from OobBuffer and exclusively use the bins
to track in-bounds range.

Changes:
 - add InRange(value) API to binhistogram, OutOfBounds(val) to Carp
 - remove bounds checks from OobBuffer Insert as this will be
   handled by the caller through the bins (so Insert is a void now).
 - Carp::UpdateInBoundsRange() to update OOB ibrange_ no longer needed
 - redirect Carp::GetInBoundsRange() to bins_ GetRange()
 - add range to OobBuffer GetPartitionedProps() API to adjust
   for removal of ibrange_
 - remove OobBuffer OutOfBounds(), SetInBoundsRange(), GetInBoundsRange()
 - InvocationPolicy::BufferInOob() uses carp_.OutOfBounds() now
 - update tests, tools usage for above