avoid unexpected use of fwrite() in msg_abort()
msg_abort is called from preload_init(), so it should not use
any functions that we preload (e.g. fwrite()).  note that gcc
 will convert simple printf/fputs into fwrite calls, so we avoid
that by using one big fprintf() before calling abort().

we need to be careful not to call functions that we
are preloading from preload_init() or we will deadlock trying
to lock init_once when we are already holding the lock for it.